+++ lib/ts/recipe/thirdparty/index.ts
@@ -110,9 +110,9 @@ export default class Wrapper {
 
     static GoogleWorkspaces = thirdPartyProviders.GoogleWorkspaces;
 
-    static Okta = thirdPartyProviders.Okta;
+    // static Okta = thirdPartyProviders.Okta;
 
-    static ActiveDirectory = thirdPartyProviders.ActiveDirectory;
+    // static ActiveDirectory = thirdPartyProviders.ActiveDirectory;
 }
 
 export let init = Wrapper.init;
@@ -149,8 +149,8 @@ export let Discord = Wrapper.Discord;
 
 export let GoogleWorkspaces = Wrapper.GoogleWorkspaces;
 
-export let Okta = Wrapper.Okta;
+// export let Okta = Wrapper.Okta;
 
-export let ActiveDirectory = Wrapper.ActiveDirectory;
+// export let ActiveDirectory = Wrapper.ActiveDirectory;
 
 export type { RecipeInterface, User, APIInterface, APIOptions, TypeProvider };