+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--   Sign in with Discord, Google workspaces, Active directory and Okta.
+-   Sign in with Discord, Google workspaces.
 -   If `getProfileInfo` throws an error, then it is sent as a FIELD_ERROR to the frontend
 
 ### Changed